requirements.txt
yfinance
pandas
numpy
streamlit
jupyter
praw
vaderSentiment
python-dotenv
requests
